Fix some timing issues with mpg video playback

+++ src/core/time.h
@@ -1,6 +1,8 @@
 #ifndef CORE_TIME_H
 #define CORE_TIME_H
 
+#include <stdint.h>
+
 /**
  * @file
  * Time tracking functions.
@@ -9,7 +11,7 @@
 /**
  * Time in millisecond-precision. Use only for time difference calculations.
  */
-typedef unsigned int time_millis;
+typedef uint64_t time_millis;
 
 /**
  * Gets the current time
